# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LeadsPay AI Actions API",
    "description": "API oficial da LeadsPay para GPT Actions e agentes compatíveis com OpenAPI. Usa a mesma API Key segura do servidor MCP.",
    "version": "1.0.0",
    "contact": {
      "name": "LeadsPay",
      "url": "https://www.techify.sbs"
    }
  },
  "servers": [
    {
      "url": "https://www.techify.sbs/api/mcp/v1",
      "description": "Produção LeadsPay"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
...
    "/coupons": {
      "get": {
        "operationId": "listCoupons",
        "summary": "Lista cupons da empresa autenticada",
        "responses": {
          "200": {
            "description": "Cupons da empresa"
          },
          "403": {
            "description": "Requer perfil Empresa aprovado"
          }
        }
      },
      "post": {
        "operationId": "createCoupon",
        "summary": "Cria ou atualiza um cupom da empresa autenticada",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCouponParams"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Cupom salvo"
          },
          "400": {
            "description": "Cupom inválido"
          },
          "403": {
            "description": "Requer perfil Empresa aprovado"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "LeadsPay API Key",
        "description": "Use a chave lp_live_ gerada em Assistentes de IA & MCP. O valor completo é exibido somente na criação."
      }
    },
    "schemas": {
...
      "CreateCouponParams": {
        "type": "object",
        "required": [
          "code",
          "discount"
        ],
        "properties": {
          "code": {
            "type": "string",
            "minLength": 3,
            "maxLength": 40
          },
          "discount": {
            "type": "number",
            "exclusiveMinimum": 0
          },
          "discountType": {
            "type": "string",
            "enum": [
              "percentage",
              "fixed"
            ],
            "default": "percentage"
          },
          "maxUses": {
            "type": "integer",
            "minimum": 0,
            "default": 100
          },
          "expiresAt": {
            "type": "string",
            "description": "Data ISO ou DD/MM/AAAA"
          },
          "planId": {
            "type": "string",
            "description": "ID da oferta ou all"
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}